backend/requirements.txt
# ============================================
# OfflineRAG Backend Dependencies
# ============================================

# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
websockets==12.0

# Async Support
aiofiles==23.2.1
aiohttp==3.9.3
httpx==0.26.0

# Data Validation
pydantic==2.6.1
pydantic-settings==2.1.0

# Vector Database
chromadb==0.4.22
hnswlib==0.8.0

# Embeddings
sentence-transformers==2.3.1
torch==2.2.0
transformers==4.37.2

# Document Processing
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.0
PyPDF2==3.0.1
pdfplumber==0.10.3
pymupdf==1.23.22
python-pptx==0.6.23
chardet==5.2.0

# OCR
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0
easyocr==1.7.1

# Audio/Video Processing
openai-whisper==20231117
pydub==0.25.1
moviepy==1.0.3
soundfile==0.12.1
librosa==0.10.1

# Text-to-Speech
TTS==0.22.0
piper-tts==1.2.0

# Text Processing
tiktoken==0.5.2
langdetect==1.0.9
beautifulsoup4==4.12.3
markdownify==0.11.6
python-magic-bin==0.4.14  # Windows
unstructured==0.12.4

# LLM Integration
ollama==0.1.6
llama-cpp-python==0.2.50

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utilities
numpy==1.26.4
scipy==1.12.0
scikit-learn==1.4.0
tenacity==8.2.3
python-dotenv==1.0.1
loguru==0.7.2
tqdm==4.66.1
watchdog==4.0.0
psutil==5.9.8

# Security
cryptography==42.0.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
httpx==0.26.0
